// src/app/models/Educacion.ts
export interface Educacion{
    id?: any;
    school: string;
    img: string;
    title: string;
    schoolStart: string;
    schoolEnd: string;

    

    // constructor (school: string, img:string, title : string, schoolStart: string, schoolEnd:string) {
    
       
    //     this.school = school;
    //     this.img = img;
    //     this.title = title;
    //     this.schoolStart= schoolStart;
    //     this.schoolEnd = schoolEnd;
    // }


}